requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
python-docx==0.8.11
pdfminer.six==20221105
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
